Allow beeline to match Elixir's OTP naming
GenServer's in elixir can be one of several type structures.
https://hexdocs.pm/elixir/GenServer.html#t:name/0

The current implementation of Beeline only works with atoms (which match
to OTP's `{:local, name}` for named processes, but this limits the
places BeeLine can be used as it either leaks atoms or prevents from
having more ant one GenServer of the same name running at a single time.
By updating the naming to include the global and via options the user
can decide how to control this instead of the library.